refactor/merge admin and content layouts
