+++ app/(content)/admin/head.tsx
@@ -1,4 +1,4 @@
-import { HeadTags } from '../../components/HeadTags';
+import { HeadTags } from '../../../components/HeadTags';
 
 export default function Head() {
   return <HeadTags title="Panel admina" robots={false} />;
