+++ app/(content)/admin/page.tsx
